Fix deleting of flows and tickets which are referenced by messages

+++ temba/archives/models.py
@@ -103,9 +103,10 @@ class Archive(models.Model):
         paginator = s3_client.get_paginator("list_objects_v2")
         for page in paginator.paginate(Bucket=settings.ARCHIVE_BUCKET, Prefix=f"{org.id}/"):
             archive_objs = page.get("Contents", [])
-            s3_client.delete_objects(
-                Bucket=settings.ARCHIVE_BUCKET, Delete={"Objects": [{"Key": o["Key"]} for o in archive_objs]}
-            )
+            if archive_objs:
+                s3_client.delete_objects(
+                    Bucket=settings.ARCHIVE_BUCKET, Delete={"Objects": [{"Key": o["Key"]} for o in archive_objs]}
+                )
 
     def filename(self):
         url_parts = urlparse(self.url)
